changed double to long double

+++ dimensions/2/prac/inc/fft_post_comp.hpp
@@ -11,28 +11,28 @@
 #define TARGET_VALUE  255.
 
 //Scale the data by the provided value
-void scale(double *data, int size, double valu);
+void scale(long double *data, int size, long double valu);
 
 //Arrange the data so that the biggest member (in module) is
 //in the middle, all of the members are circulary shifted 
 //such that such an arrangement is achieved
 //In practice, the two halves get swapped
-void nyquist_arrange(double *data, int size);
+void nyquist_arrange(long double *data, int size);
 
 //Works properly only after the "nyquist_arrange" function was called
 //on the input 'data'
-void set_max_vals(double *data, int size, double *vals);
+void set_max_vals(long double *data, int size, long double *vals);
 
 //Compute the constants from the given max values, the max values
 //extracted by the previous function
-void const_from_max(double target, double *vals);
+void const_from_max(long double target, long double *vals);
 
 //Save modulus to the real part, the imaginary one gets nullified in the process
 //The mirror function is not needed
-void modulus_in_real(double *data, int size); 
+void modulus_in_real(long double *data, int size); 
 
 //Logarithmically normalize the real parts by the provided value 
 //The mirror function is not needed
-void log_norm_real(double *data, int size, double *vals);
+void log_norm_real(long double *data, int size, long double *vals);
 
 #endif//_FFT_POST_COMP_H_
